move simd code back to `vortex-array`
